+++ cpp/arcae/configuration.cc
@@ -1,39 +1,31 @@
 #include "arcae/configuration.h"
 
+#include <string>
 
 namespace arcae {
 
 arrow::Result<std::string> Configuration::Get(const std::string & key) const {
-    auto it = kvmap_.find(key);
-
-    if(it == kvmap_.end()) {
-        return arrow::Status::KeyError(key);
-    } else {
+    if(auto it = kvmap_.find(key); it != kvmap_.end()) {
         return it->second;
     }
+    return arrow::Status::KeyError(key);
 }
 
 
 std::string Configuration::GetDefault(const std::string & key, std::string default_value) const {
-    auto it = kvmap_.find(key);
-
-    if(it == kvmap_.end()) {
-        return std::move(default_value);
-    } else {
+    if(auto it = kvmap_.find(key); it != kvmap_.end()) {
         return it->second;
     }
+    return default_value;
 }
 
 
 arrow::Result<bool> Configuration::Delete(const std::string & key) {
-    auto it = kvmap_.find(key);
-
-    if(it == kvmap_.end()) {
-        return arrow::Status::KeyError(key);
-    } else {
+    if(auto it = kvmap_.find(key); it != kvmap_.end()) {
         kvmap_.erase(it);
         return true;
     }
+    return arrow::Status::KeyError(key);
 }
 
 } // namespace arcae