+++ cpp/tests/test_utils.h
@@ -1,5 +1,12 @@
+#ifndef ARCAE_TEST_UTILS_H
+#define ARCAE_TEST_UTILS_H
+
+#include <string>
 namespace arcae {
+
 
 std::string hexuuid(std::size_t n=8);
 
-}
+}  // namespace
+
+#endif //  ARCAE_TEST_UTILS_H